Implemented Promise as a return value
