+++ package.json
@@ -5,7 +5,7 @@
   "main": "build/src/faussaire.js",
   "scripts": {
     "lint": "eslint . && echo 'Lint finished...\n'",
-    "test": "jasmine",
+    "test": "jest --watch",
     "start": "webpack-dev-server --content-base build/",
     "build": "babel src/* --out-dir build",
     "update": "updtr"
@@ -19,7 +19,7 @@
     "network",
     "request"
   ],
-  "author": "Anthony Benkhebbab",
+  "author": "Anthony Benkhebbab <rewieer@gmail.com>",
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/Rewieer/faussaire/issues"
@@ -34,8 +34,10 @@
     "babel-plugin-transform-object-rest-spread": "^6.16.0",
     "babel-preset-es2015": "^6.16.0",
     "babel-preset-node6": "^11.0.0",
+    "babel-preset-stage-3": "^6.17.0",
     "eslint": "^3.6.1",
     "jasmine": "^2.5.2",
+    "jest": "^18.0.0",
     "updtr": "^0.2.1",
     "webpack": "^1.13.2",
     "webpack-dev-server": "^1.16.1"
