--- spec/support/jasmine.json
@@ -1,12 +1,0 @@
-{
-  "spec_dir": "spec",
-  "spec_files": [
-    "**/*-test.js"
-  ],
-  "helpers": [
-    "helpers/**/*.js",
-    "../node_modules/babel-register/lib/node.js"
-  ],
-  "stopSpecOnExpectationFailure": false,
-  "random": false
-}
