+++ src/faussaire.js
@@ -1,85 +1,7 @@
-import responseFactory from './response';
-import routeFactory from './route';
-import controllerFactory from './controller';
-
-/**
- * Regex to match urls ending with "?args=value" and to make it
- * match precise routes.
- * See https://github.com/Rewieer/faussaire/issues/1
- *
- * @type {string}
- */
-const URLArgsRegex = "((\\?)([^=]+)(=(.+))?)?$";
-
-/**
- * Return true if the url is matching the route
- *
- * @param route
- * @param url
- * @returns {boolean}
- */
-const isMatching = (route, url) => {
-  const urlRegex = route.replace(/{(\w)+}/g, "([^/]+)") + URLArgsRegex;
-  return new RegExp(urlRegex).test(url);
-};
-
-/**
- * Extract all the arguments from the parameters, following the ? in the URL
- * @param url
- * @returns {{}}
- */
-const extractURLArgs = (url) => {
-  let str = url.split('?'),
-      obj = {}
-  ;
-
-  if(str[1]){
-    let pairs = str[1].split('&');
-
-    if(pairs.length === 0){
-      return obj;
-    }
-
-    [].slice.call(pairs).forEach(function(pair){
-      let keyValue = pair.split("=");
-      obj[keyValue[0]] = keyValue[1];
-    });
-
-    return obj;
-  }
-
-  return {};
-};
-
-/**
- * Extract routing parameters
- *
- * From template such as http://url.com/post/{id}
- * with url such as http://url.com/post/3
- * This function will match ID with the value 3.
- *
- * @param template
- * @param url
- * @returns {{}}
- */
-const extractRouteParameters = function(template, url){
-  let keys = [];
-
-  let urlRegex = template.replace(/{(\w)+}/g, function(arg){
-    keys.push(arg.substr(1, arg.length - 2));
-    return "([^?]+)";
-  });
-
-  let regex = new RegExp(urlRegex);
-  let routeArgs = regex.exec(url);
-
-  let obj = {};
-  [].slice.call(routeArgs, 1).forEach(function(t, i){
-    obj[keys[i]] = t;
-  });
-
-  return obj;
-};
+import responseFactory from './request/response';
+import routeFactory from './request/route';
+import controllerFactory from './request/controller';
+import { isMatching, extractURLArgs, extractRouteParameters } from './stringUtil';
 
 /**
  * Create a faussaire instance
@@ -131,62 +53,64 @@ const createFaussaire = () => {
      * @param url
      * @param method
      * @param requestBody
-     * @returns response
+     * @returns Promise
      */
     fetch: (url, method, requestBody = {}) => {
-      const matchingRoute = _routes.find(r =>
-        isMatching(r.template, url) && r.methods.indexOf(method.toUpperCase()) > -1
-      );
+      return new Promise((accept, reject) => {
+        const matchingRoute = _routes.find(r =>
+          isMatching(r.template, url) && r.methods.indexOf(method.toUpperCase()) > -1
+        );
 
-      if(!matchingRoute) {
-        throw throwError({
-          response: _onNotFoundError
-        });
-      }
+        if(!matchingRoute) {
+          reject(throwError({
+            response: _onNotFoundError
+          }));
+        }
 
-      let query   = [],
+        let query   = [],
           request = [],
           route   = extractRouteParameters(matchingRoute.template, url)
-        ;
+          ;
 
-      // In GET methods, there's no need to read request's body
-      // If there is a requestBody in the fetch, the user still probably
-      // Wants them to be considered as query parameters
-      if(method === "GET"){
-        query = Object.assign({}, extractURLArgs(url), requestBody.params);
-        request = [];
-      } else {
-        query = Object.assign({}, extractURLArgs(url), requestBody.params);
-        request = requestBody.data;
-      }
-
-      const params = {
-        query,
-        request,
-        route
-      };
-
-      // Object holding data about the process
-      const options = {
-        method
-      };
-
-      if(typeof matchingRoute.controller.authenticate === 'function'){
-        const token = matchingRoute.controller.authenticate(params, options);
-
-        if(typeof token !== 'undefined'){
-          options.token = token;
+        // In GET methods, there's no need to read request's body
+        // If there is a requestBody in the fetch, the user still probably
+        // Wants them to be considered as query parameters
+        if(method === "GET"){
+          query = Object.assign({}, extractURLArgs(url), requestBody.params);
+          request = [];
+        } else {
+          query = Object.assign({}, extractURLArgs(url), requestBody.params);
+          request = requestBody.data;
         }
-      }
 
-      const response = matchingRoute.controller.run(params, options);
-      if(response.status >= 400){
-        throw throwError({
-          response
-        });
-      }
+        const params = {
+          query,
+          request,
+          route
+        };
 
-      return response;
+        // Object holding data about the process
+        const options = {
+          method
+        };
+
+        if(typeof matchingRoute.controller.authenticate === 'function'){
+          const token = matchingRoute.controller.authenticate(params, options);
+
+          if(typeof token !== 'undefined'){
+            options.token = token;
+          }
+        }
+
+        const response = matchingRoute.controller.run(params, options);
+        if(response.status >= 400){
+          reject(throwError({
+            response
+          }));
+        }
+
+        return accept(response);
+      });
     },
 
     /**
